Adding constant time steps

+++ Project 3/src/com/hw3/sketcher/Movable.java
@@ -5,5 +5,6 @@ public interface Movable {
 	public void step();
 	public void stop();
 	public void reset();
-	public float[] getSpeed();
+	public double[] getSpeed();
+
 }
